fix: deployment scripts when using other output formats in profile configuration
By default the AWS CLI uses JSON as a output format. But when you have set your named profiles to for example YAML, the deployment scripts will not work.
By supplying the `--output json` overwrite to the commands that use `jq` we ensure that the input of `jq` is indeed JSON.